Use Double.isNaN instead of Float.isNaN

+++ src/ru/nsu/ccfit/zuev/osu/game/Slider.java
@@ -309,7 +309,7 @@ public class Slider extends GameObject {
         // Try to fix incorrect slider tick count bug
         tickInterval = timing.getBeatLength() * speedMultiplier;
         int tickCount = (int) (maxTime * GameHelper.getTickRate() / tickInterval);
-        if(Float.isNaN((float) tickInterval) || tickInterval < GameHelper.getSliderTickLength() / 1000){
+        if(Double.isNaN(tickInterval) || tickInterval < GameHelper.getSliderTickLength() / 1000){
             tickCount = 0;
         }
         if ((maxTime * GameHelper.getTickRate() / tickInterval)
